organize file download unit tests into different files
